chore: pre-flight fixes — split stacked VM tests, key-path lint (tc-vay6)
